+++ src/main/java/com/ApiRestStock/CRUD/Finanzas/Compra/DTOs/DetalleCompraResponse.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public record DetalleCompraResponse(
         Long id,
-        Integer cantidad,
+        BigDecimal cantidad,
         BigDecimal precioUnitario,
         String nombreProducto,
         Long productoId    // solo el id, no todo el Producto
