+++ src/main/java/com/ApiRestStock/CRUD/ventas/DTOs/DetalleVentaResponse.java
@@ -5,6 +5,6 @@ import java.math.BigDecimal;
 public record DetalleVentaResponse(
     Long id,
     String nombreProducto,
-    Integer cantidad,
+    BigDecimal cantidad,
     BigDecimal precioUnitario
 ) {}
